requirements.txt
django
gunicorn
psycopg2-binary==2.9.5
djangorestframework
requests
pillow
